Make it possible to load TX history using Native mode * Replace TransferHistoryBuilder from swap.rs to history.rs * Add TransferHistoryBuilder::build_tx_idents() that is used in request_tx_history()
